finalizacao do trabalho

+++ 05 - JSP/atvJSP/src/main/webapp/removerPostagem.jsp
@@ -11,31 +11,48 @@
 <body>
 	
 		
-	<%		
+	<%
+	
+	int permissao = 0;
+	String session_u_name = (String)session.getAttribute("user");
+	String[] dados = null;
+	boolean estaLogado = false;
+	if(session_u_name != null)
+	{
+	dados = session_u_name.split(",");
+	permissao = Integer.parseInt(dados[1]);                  
+	   estaLogado = true;
+	}
+
+	if(permissao == 0){
+		response.sendError(401, "Você precisa ser um administrador para realizar esta ação.");
+	}
+	else
+	{
 		int codigo = Integer.parseInt(request.getParameter("codigo"));
 		
 		String deleteC = "DELETE FROM comentarios WHERE codigo_noticia = ?";
 		
-		//Efetuar a conexao
+		
 		Conexao c = new Conexao();
 		PreparedStatement pstmt = c.efetuarConexao().prepareStatement(deleteC);
 		pstmt.setInt(1,codigo);
-		//Executar a remoção
+		
 		pstmt.execute();
 
-		// Comando SQL
+
 		String sql = "DELETE FROM postagens WHERE codigo = ?";
 			
-		//PreparedStatement
+		
 		PreparedStatement pstmtP = c.efetuarConexao().prepareStatement(sql);
 		pstmtP.setInt(1,codigo);
 				
-		//Executar a remoção
+		
 		pstmtP.execute();
 				
-		//Redirecionamento
+		
 		response.sendRedirect("index.jsp");
-
+	}
 
 	%>
